Move KeyDumpSigner to Tests\ namespace
