+++ phpstan.neon.dist
@@ -3,21 +3,3 @@ parameters:
     paths:
         - src
         - tests
-
-    ignoreErrors:
-        - """
-            #^Call to deprecated method fromEmptyData\\(\\) of class Lcobucci\\\\JWT\\\\Token\\\\Signature:
-            Deprecated since v4\\.3$#
-        """
-        - """
-            #^Call to deprecated method forUnsecuredSigner\\(\\) of class Lcobucci\\\\JWT\\\\Configuration:
-            Deprecated since v4\\.3$#
-        """
-        - """
-            #^Call to deprecated method empty\\(\\) of class Lcobucci\\\\JWT\\\\Signer\\\\Key\\\\InMemory:
-            Deprecated since v4\\.3$#
-        """
-        - """
-            #^.+ of deprecated class Lcobucci\\\\JWT\\\\Signer\\\\None:
-            Deprecated since v4\\.3$#
-        """
